v0.2.0: Submitting log proof works
