+++ README.md
@@ -6,7 +6,7 @@ The aim of this project is to provide a mechanism for enabling smart contracts t
 *Note:* Ethbase is currently very early stage, use at own risk!
 
 ## Usage
-In the given [example](contracts/example), we have an `Emitter` contract, which emits `Transfer(uint256)` when its method is called. We want our `Subscriber` contract to update its state, whenever `Transfer` is emitted. This can be done, by subscribing to the registry, and specifying events and their callback methods. Afterwards, any user who's running an Ethereum client, upon seeing a `Transfer` event, can call `invoke` on `Registry`, which would in turn invoke `Subscriber`'s specified method. Of course, this manual invocation is not practical. In future, there could be a daemon which listens to events, invoking methods automatically, and smart contracts that subscribe would pay a small fee per event invocation.
+In the given [example](contracts/example), we have an `Emitter` contract, which emits `Transfer(uint256)` when its method is called. We want our `Subscriber` contract to update its state, whenever `Transfer` is emitted. This can be done, by subscribing to the registry, and specifying events and their callback methods. Afterwards, any user who's running an Ethereum client, upon seeing a `Transfer` event, can call `submitLog` on `Ethbase`, which would in turn invoke `Subscriber`'s specified method. Of course, this manual invocation is not practical. In future, there could be a daemon which listens to events, invoking methods automatically, and smart contracts that subscribe would pay a small fee per event invocation.
 
 ## Contribute
 Most of the discussions are happening in the [issues](https://github.com/planet-ethereum/ethbase/issues) and [pull requests](https://github.com/planet-ethereum/ethbase/pulls), Suggestions, contributions and critisisms are more than welcome, join in.
